pyproject.toml
[tool.poetry]
name = "notification-services"
version = "1"
description = "packages"
authors = ["Рифат Хасанов <Hidex-morfeyka@mail.ru>"]
readme = "README.md"
packages = [{include = "notification_services"}]

[tool.poetry.dependencies]
python = "^3.10"
amqp = "5.1.1"
asgiref = "3.5.2"
attrs = "22.1.0"
billiard = "3.6.4.0"
cachecontrol = "0.12.11"
celery = "5.2.7"
certifi = "2022.12.7"
charset-normalizer = "2.1.1"
cleo = "2.0.1"
click = "8.1.3"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.2.0"
colorama = "0.4.6"
coreapi = "2.3.3"
coreschema = "0.0.4"
crashtest = "0.4.1"
distlib = "0.3.6"
django = "4.1.4"
django-rest-framework = "0.1.0"
djangorestframework = "3.14.0"
drf-yasg = "1.21.4"
dulwich = "0.20.50"
filelock = "3.8.2"
html5lib = "1.1"
idna = "3.4"
importlib-metadata = "5.1.0"
inflection = "0.5.1"
itypes = "1.2.0"
jaraco-classes = "3.2.3"
jinja2 = "3.1.2"
jsonschema = "4.17.3"
keyring = "23.11.0"
kombu = "5.2.4"
lockfile = "0.12.2"
markupsafe = "2.1.1"
more-itertools = "9.0.0"
msgpack = "1.0.4"
oauthlib = "3.2.2"
packaging = "22.0"
pexpect = "4.8.0"
pkginfo = "1.9.2"
platformdirs = "2.6.0"
poetry = "1.3.1"
poetry-core = "1.4.0"
poetry-plugin-export = "1.2.0"
prompt-toolkit = "3.0.36"
psycopg2 = "2.9.5"
ptyprocess = "0.7.0"
pyrsistent = "0.19.2"
python-dotenv = "0.21.0"
pytz = "2022.6"
pywin32-ctypes = "0.2.0"
rapidfuzz = "2.13.6"
requests = "2.28.1"
requests-oauthlib = "1.3.1"
requests-toolbelt = "0.10.1"
ruamel-yaml = "0.17.21"
ruamel-yaml-clib = "0.2.7"
shellingham = "1.5.0"
six = "1.16.0"
sqlparse = "0.4.3"
tomli = "2.0.1"
tomlkit = "0.11.6"
trove-classifiers = "2022.12.1"
tzdata = "2022.7"
uritemplate = "4.1.1"
urllib3 = "1.26.13"
vine = "5.0.0"
virtualenv = "20.17.1"
wcwidth = "0.2.5"
webencodings = "0.5.1"
zipp = "3.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
